Added Level 2
#16  I created the second level and set up all the managers and UI so it will work.

+++ JustinWorld/Assets/Scripts/Player/PlayerController.cs
@@ -11,7 +11,7 @@ public class PlayerController : MonoBehaviour {
     protected Collider2D col;
     protected float direction;
     public int speed;
-    public int jumpForce;
+    public float jumpForce;
     public Transform feet;
     public Transform cameraTarget;
     private Collider2D lastTrigger;
